Change data type

+++ panel.go
@@ -164,11 +164,11 @@ type (
 		Decimals        int      `json:"decimals"`
 		Format          string   `json:"format"`
 		Gauge           struct {
-			MaxValue         int  `json:"maxValue"`
-			MinValue         int  `json:"minValue"`
-			Show             bool `json:"show"`
-			ThresholdLabels  bool `json:"thresholdLabels"`
-			ThresholdMarkers bool `json:"thresholdMarkers"`
+			MaxValue         float32 `json:"maxValue"`
+			MinValue         float32 `json:"minValue"`
+			Show             bool    `json:"show"`
+			ThresholdLabels  bool    `json:"thresholdLabels"`
+			ThresholdMarkers bool    `json:"thresholdMarkers"`
 		} `json:"gauge,omitempty"`
 		MappingType     *uint       `json:"mappingType,omitempty"`
 		MappingTypes    []*MapType  `json:"mappingTypes,omitempty"`
